Integrates local storage caching using react native async storage.
In this commit, we refactor the old StorageManager library into the
current project replacing the storage api with react natives
AsyncStorage. We also change a couple of the data type responses and
make the code more consistent. For the stop and service components
we fetch the relevant data on the screen. We also add the Firebase
library however we have not implemented such functionality.