Creates MetlinkListItem component for tidy code.
In this commit we create a couple notable changes.
The first of which is we create a generic list
component for metlink stops and services. To support
this we also support a new dynamic navigation method
to navigate to metlink stops and services from the
current path. Sadly we had to bring the scroll bar to
a more parent component as it was taking too long to
figure out the styles dynamically (but also fixed
values).