Switches to csv() library, adds RN toasts and adds share popup functionality.
In this commit, we implment a couple significant changes. This has all been
put in a single commit as we are testing that all of these different components
are working on mobile. Unfortunately we had to switch CSV library too
(Papaparse CSV was really nice) as what we were using depended on some node.js
functions that aren't available on mobile. We also introduce toasts that popup
on desktops where the share api does not work. This library is behaving funny,
and Webstorm can't detect the import however it appears to work just fine. Have
tsxignored the issue for the time being as it appears to work flawlessly for
now.