Implements TwitterFeedContainer component.
In this commit we implement the twitter feed component. Unfortunately this
was a bit more awkward than initially expected as classes aren't as easily
accessible with React Native than they were with Ionic. Instead we chose
to grab the parent of the <a/> link which we can cheekily add a class to.

IMPORTANT: Due to the Papa Parse library not being valid for Mobile, I have
not been able to test this on Android yet.